requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiofiles=23.2.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.7.4=hbcca054_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.2=py312h7070661_0
decorator=5.1.1=pyhd8ed1ab_0
dnspython=2.6.1=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
email-validator=2.2.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.111.1=pypi_0
fastapi-cli=0.0.4=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.15.4=pypi_0
fonttools=4.53.1=pypi_0
fsspec=2024.6.1=pypi_0
gradio=4.38.1=pypi_0
gradio-client=1.1.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.24.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.0.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=8.0.0=hd8ed1ab_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py312h7900ff3_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.3=h659f571_0
ld_impl_linux-64=2.40=hf3520f5_7
libedit=3.1.20191231=he28a2e2_2
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=14.1.0=h77fa898_0
libgomp=14.1.0=h77fa898_0
libnsl=2.0.1=hd590300_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.46.0=hde9e2c9_0
libstdcxx-ng=14.1.0=hc0a3c3a_0
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=h4ab18f5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h59595ed_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.3=pypi_0
numpy=2.0.0=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.82=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.3.1=h4bc722e_2
orjson=3.10.6=pypi_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=pypi_0
pip=24.0=pyhd8ed1ab_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
psutil=6.0.0=py312h9a8786e_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
python=3.12.4=h194c7f8_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.12=4_cp312
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=py312h8fd38d8_0
readline=8.2=h8228510_1
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.19.0=pypi_0
ruff=0.5.3=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=71.0.3=pyhd8ed1ab_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
sympy=1.13.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.3.1=pypi_0
torchvision=0.18.1=pypi_0
tornado=6.4.1=py312h9a8786e_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.12.3=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.1=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.22.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websockets=11.0.3=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h166bdaf_0
zeromq=4.3.5=h75354e8_4
zipp=3.19.2=pyhd8ed1ab_0
